# core/API-swagger.yaml
openapi: 3.1.0
info:
  title: SupremeAI 2.0
  version: 0.1.0
paths:
...
  /api/v1/meta-ai/breed:
    post:
      tags:
      - layer-6-meta-ai
      summary: Run Breeding Cycle
      description: "Run a full breeding cycle: select parents \u2192 breed \u2192\
        \ evaluate \u2192 promote."
      operationId: run_breeding_cycle_api_v1_meta_ai_breed_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api__routes__meta_ai__BreedRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/meta-ai/pool:
    get:
      tags:
      - layer-6-meta-ai
      summary: List Breeding Pools
      description: List breeding pools.
      operationId: list_breeding_pools_api_v1_meta_ai_pool_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - layer-6-meta-ai
      summary: Create Breeding Pool
      description: Create or update breeding pool.
      operationId: create_breeding_pool_api_v1_meta_ai_pool_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolCreateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Breeding Pool Api V1 Meta Ai Pool Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/meta-ai/weakest-links:
    get:
      tags:
      - layer-6-meta-ai
      summary: Get Weakest Links
      description: Identify and list weakest links.
      operationId: get_weakest_links_api_v1_meta_ai_weakest_links_get
      security:
      - HTTPBearer: []
      parameters:
      - name: agent_names
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Names
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeakestLinkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/meta-ai/weakest-links/{id}/ack:
    post:
      tags:
      - layer-6-meta-ai
      summary: Acknowledge Weakest Link
      description: Acknowledge weakest link report.
      operationId: acknowledge_weakest_link_api_v1_meta_ai_weakest_links__id__ack_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Acknowledge Weakest Link Api V1 Meta Ai Weakest Links  Id  Ack
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/meta-ai/top-performers:
    get:
      tags:
      - layer-6-meta-ai
      summary: Get Top Performers
      description: Get top-N performing agents.
      operationId: get_top_performers_api_v1_meta_ai_top_performers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 5
          title: Limit
      - name: lookback_hours
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lookback Hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopPerformerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BreedResponse:
      properties:
        success:
          type: boolean
          title: Success
        offspring:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Offspring
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: BreedResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PoolCreateRequest:
      properties:
        pool_name:
          type: string
          title: Pool Name
        agent_names:
          items:
            type: string
          type: array
          title: Agent Names
        min_fitness_threshold:
          type: number
          title: Min Fitness Threshold
          default: 0.6
        max_pool_size:
          type: integer
          title: Max Pool Size
          default: 20
      type: object
      required:
      - pool_name
      - agent_names
      title: PoolCreateRequest
    PoolResponse:
      properties:
        pools:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Pools
      type: object
      required:
      - pools
      title: PoolResponse
...
    TopPerformerResponse:
      properties:
        top_performers:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Top Performers
      type: object
      required:
      - top_performers
      title: TopPerformerResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WeakestLinkResponse:
      properties:
        reports:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Reports
        generated_at:
          type: string
          title: Generated At
      type: object
      required:
      - reports
      - generated_at
      title: WeakestLinkResponse
...
    api__routes__meta_ai__BreedRequest:
      properties:
        pool_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Pool Name
          description: Specific breeding pool to use
        parent_a:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent A
          description: Explicit parent A name
        parent_b:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent B
          description: Explicit parent B name
        offspring_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Offspring Name
          description: Desired offspring name
      type: object
      title: BreedRequest
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login